ipsec: Generate Generate to-<somewhere> trace for ipsec packets
When ipsec encrypted packets come to the from-host program (bpf_host),
in case of the native routing mode, it will be handled by
do_netdev_encrypt, then redirected to the network-facing device when the
fib_lookup is available. Otherwise, it will be sent to stack for
routing. There are three possible packet paths.

1. fib_lookup available + to-netdev program attached to the dest device

In this case, to-netdev program may drop packet with host firewall. So,
we shouldn't generate TRACE_TO_NETWORK trace before seeing the verdict.
Thus, trace should be generated from to-netdev program.

2. fib_lookup available + to-netdev program is not attached

In this case, packet will go to the network without involving BPF
program. Thus, we should generate trace in the context of from-host
program.

3. Packet goes to stack and route

Unlike above two, packet goes to stack. In this case, TRACE_TO_STACK
should be generated. However, we don't have to generate the trace from
from-host program, because if we do CTX_ACT_OK from from-host program,
the packet pass through veth and hooked into to-host program
immediately. In to-host program, there is a TRACE_TO_STACK.

Trace on source node before fix

```
<- endpoint ... EchoRequest
-> stack ... EchoRequest
<- host encrypted ...
```

Trace on the source node after fix

```
<- endpoint ... EchoRequest
-> stack ... EchoRequest
<- host encrypted ...
-> network encryped ...
```

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>