clustermesh: Add an infrastructure to manage per-cluster configuration
When the cluster configuration file is updated, currently we are just
passing the path to the file to remoteCluster object. The path is passed
to the etcd client and etcd client will read the contents of the file.

Since we started to use cluster configuration file to store additional
configuration like has-overlapping-pod-cidr flag, we need an
infrastructure to manage them.

This commit adds a new struct
remoteClusterConfig which contains the mapping of yaml fields **other
than** etcd-related one. It can be made from configuration file by
parseRemoteClusterConfig().

Currently, it contains hasOverlappingPodCIDR
which is mapped to the has-overlapping-pod-cidr flag. We can add any
per-cluster configuration in the future to here.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>