clustermesh: Add ClusterID to Service in kvstore
Like what we do with Node, include ClusterID to the Service object in
kvstore. This ClusterID will be used for identifying service backends
when remote cluster has an overlapping PodCIDR from local cluster's
perspective.

To achieve that, we extend the k8s -> kvstore service synchronization
logic and add ClusterID information to service object while we
converting it from k8s to kvstore representation. To teach ClusterID to
the service synchronizer, we needed to extend the ServiceSyncConfiguration
interface with a new method LocalClusterID(). This affecs to all of the
components implement this interface. That's why this commit contains the
change to DaemonConfig as well.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>