clustermesh: Update services_test to use a new kvstore service format
Update existing services_test with a new service format on kvstore which
contains ClusterID. Now prepareServiceUpdate() takes ClusterID as an
argument. Existing tests are updated to explicitely set ClusterID, but
it shouldn't break the tests because we don't set hasOverlappingPodCIDR
for clusters and the behavior should be the same as before in that case.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>