clustermesh: Introduce cluster-aware addressing
Introduce a new addressing scheme which identifies the network endpoints
with IP address + ClusterID. With this special addressing scheme, we can
distinguish two network endpoints that have a same IP address, but
located in the different cluster.

This commit introduces a new type AddrCluster to express the address as
well as the string notation of it. Please see ParseAddrCluster function
for more details.

Signed-off-by: Yutaro Hayakawa <yutaro.hayakawa@isovalent.com>